+++ CHANGELOG.md
@@ -10,8 +10,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add download and sign contracts page on the learner dashboard.
 - Display user full_name instead of username if it exists
-- Add a footer on enrollment's item in the learner dashboard. It give the
+- Add a footer on enrollment's item in the learner dashboard. It gives the
   possibility to purchase linked product or download linked certificate.
 - Add download contracts pages on the teacher dashboard.
 - Add a contract information and actions in learner dashboard order's